Real price of Iranian Rial&US Dolaar
"If you search on Google for the current inflation rate in the United States and then search on the internet for the current inflation rate in Iran, subtract the two rates from each other, divide by 100, and multiply it by the current exchange rate of the dollar to the Iranian rial. This pattern is something I continuously use and thought of delivering it to my friends in a graphical user interface format. Good luck!"